final initial package layout proposal -- no repositories, subsumed into viewmodels
