--- app/src/main/java/com/example/wildfire_fixed_imports/model/retro.kt
@@ -1,2 +1,0 @@
-package com.example.wildfire_fixed_imports.model
-
