--- app/src/main/java/com/example/wildfire_fixed_imports/view/settings/SettingsViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.wildfire_fixed_imports.view.settings
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class SettingsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
-    }
-    val text: LiveData<String> = _text
-}
