--- app/src/main/java/com/example/wildfire_fixed_imports/view/tools/ToolsFragment.kt
@@ -1,31 +1,0 @@
-package com.example.wildfire_fixed_imports.view.tools
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
-import com.example.wildfire_fixed_imports.R
-
-class ToolsFragment : Fragment() {
-
-    private lateinit var toolsViewModel: ToolsViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        toolsViewModel =
-            ViewModelProviders.of(this).get(ToolsViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_tools, container, false)
-        val textView: TextView = root.findViewById(R.id.text_tools)
-        toolsViewModel.text.observe(this, Observer {
-            textView.text = it
-        })
-        return root
-    }
-}
