--- app/src/main/java/com/example/wildfire_fixed_imports/view/tools/ToolsViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.wildfire_fixed_imports.view.tools
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ToolsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is tools Fragment"
-    }
-    val text: LiveData<String> = _text
-}
