+++ app/src/main/java/com/example/wildfire_fixed_imports/viewmodel/view_controllers/MapController.kt
@@ -1,0 +1,2 @@
+package com.example.wildfire_fixed_imports.viewmodel.view_controllers
+
