+++ app/src/main/java/com/example/wildfire_fixed_imports/viewmodel/vmclasses/SettingsViewModel.kt
@@ -1,0 +1,13 @@
+package com.example.wildfire_fixed_imports.viewmodel.vmclasses
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SettingsViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is setting Fragment"
+    }
+    val text: LiveData<String> = _text
+}
