+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -25,15 +25,15 @@
 
     <fragment
         android:id="@+id/nav_debug"
-        android:name="com.example.wildfire_fixed_imports.view.tools.ToolsFragment"
+        android:name="com.example.wildfire_fixed_imports.view.tools.DebugFragment"
         android:label="@string/menu_debug"
-        tools:layout="@layout/fragment_tools" />
+        tools:layout="@layout/fragment_debug" />
 
-    <fragment
+<!--    <fragment
         android:id="@+id/nav_share"
         android:name="com.example.wildfire_fixed_imports.view.share.ShareFragment"
         android:label="@string/menu_share"
-        tools:layout="@layout/fragment_share" />
+        tools:layout="@layout/fragment_share" />-->
 
 <!--    <fragment
         android:id="@+id/nav_send"
